requirements.txt
certifi==2019.9.11
chardet==3.0.4
idna==2.8
Pillow==6.2.1
pony==0.7.11
psycopg2==2.8.4
requests==2.22.0
six==1.12.0
urllib3==1.25.6
vk-api==11.6.1
